Refactor: Split reqResMapper into createRequest/ResponseObject
Split the reqResMapper function into two separate files. One for
creating the request object and one for creating the response object.
This is easier to read and understand.
