+++ lib/helpers/setupNetlifyFunctionForPage.js
@@ -24,7 +24,12 @@ const setupNetlifyFunctionForPage = ({ filePath, functionsPath }) => {
   });
 
   // Copy function helpers
-  ["renderNextPage.js", "reqResMapper.js"].forEach((helper) => {
+  const functionHelpers = [
+    "renderNextPage.js",
+    "createRequestObject.js",
+    "createResponseObject.js",
+  ];
+  functionHelpers.forEach((helper) => {
     copySync(join(TEMPLATES_DIR, helper), join(functionDirectory, helper), {
       overwrite: false,
       errorOnExist: true,
