+++ lib/templates/renderNextPage.js
@@ -1,6 +1,7 @@
 // Load the NextJS page
 const nextPage = require("./nextPage");
-const reqResMapper = require("./reqResMapper");
+const createRequestObject = require("./createRequestObject");
+const createResponseObject = require("./createResponseObject");
 
 // Render the Next.js page
 const renderNextPage = (event) => {
@@ -10,8 +11,10 @@ const renderNextPage = (event) => {
     // Create a Next.js-compatible request and response object
     // These mock the ClientRequest and ServerResponse classes from node http
     // See: https://nodejs.org/api/http.html
-    const callback = (_null, response) => resolve(response);
-    const { req, res } = reqResMapper(event, callback);
+    const req = createRequestObject({ event });
+    const res = createResponseObject({
+      onResEnd: (response) => resolve(response),
+    });
 
     // Check if page is a Next.js page or an API route
     const isNextPage = nextPage.render instanceof Function;
